Move release notes and genecentric api urls. Fix some imports.
