+++ src/shared/config/apiUrls/apiUrls.ts
@@ -1,6 +1,7 @@
 import * as configure from './configure';
 import * as entry from './entry';
 import * as proteinsApi from './proteinsApi';
+import * as releaseNotes from './releaseNotes';
 import * as search from './search';
 import * as results from './results';
 import * as suggester from './suggester';
@@ -9,6 +10,7 @@ export default {
   configure,
   entry,
   proteinsApi,
+  releaseNotes,
   results,
   search,
   suggester,
