+++ src/uniprotkb/config/apiUrls/apiUrls.ts
@@ -1,6 +1,7 @@
+import * as genecentric from './genecentric';
 import * as groupBy from './groupBy';
 import * as publications from './publications';
 import * as statistics from './statistics';
 import * as unisave from './unisave';
 
-export default { groupBy, publications, statistics, unisave };
+export default { genecentric, groupBy, publications, statistics, unisave };
