+++ src/uniprotkb/config/apiUrls/genecentric.ts
@@ -1,0 +1,10 @@
+import joinUrl from 'url-join';
+
+import { stringifyUrl } from '../../../shared/utils/url';
+
+import { apiPrefix } from '../../../shared/config/apiUrls/apiPrefix';
+
+export const search = (accession: string) =>
+  stringifyUrl(joinUrl(apiPrefix, 'genecentric/search'), {
+    query: `accession:${accession}`,
+  });
